+++ src/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: rwallier <rwallier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/12 19:27:16 by rwallier          #+#    #+#             */
-/*   Updated: 2022/05/12 19:55:23 by rwallier         ###   ########.fr       */
+/*   Updated: 2022/05/13 13:39:46 by rwallier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strtrim(char const *s1, char const *set)
 {
-	char			*ret;
-	unsigned int	start;
-	size_t			end;
+	char	*ret;
+	size_t	start;
+	size_t	end;
 
 	start = 0;
 	while (s1[start] && ft_strchr(set, s1[start]))
